\name{codco-package}
\alias{codco-package}
\alias{codco}
\docType{package}
\title{
\packageTitle{codco}
}
\description{
\packageDescription{codco}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{codco}
\packageIndices{codco}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{codco}

Maintainer: \packageMaintainer{codco}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
